fix: rename file RunsTool.ts → ExecutionsTool.ts, add legacy registry aliases
- Rename file to match exported class name (ExecutionsTool)
- Register legacy tool names (propose_workflow, propose_agent, runs) as
  aliases in the registry so old agent configs still resolve correctly
- Include legacy names in DELEGATION_TOOLS filter to prevent nesting
  bypass via old config names
